* [core]: robuster CGATS DATA_FORMAT parsing
Allow for omitting the DATA_FORMAT section and reuse that across data sections
until they are appearing newly. Example is a Monaco CGATS file.
While the CGATS style is not perfect, keep relaxed about such errors and
report.

The "goto" statement came previously before the increase of the nFelder_
member. Thus there where the last data fields counter in the various value
lists and the increased nFelder_ on the other side. That leaded to crashes.