* [core]: use the Oy_Fl_Group widget from Oyranos
MyFl_Double_Window is derived from Oy_Fl_Window_Base and put a call to that
class into its own handle() function.

The editing colour space must be set of Oy_Fl_Group. Input colours
need to be in that editing space.

The 3D status bar and the report window are colour managed this way.
It runs on CPU and works on multi monitors in native device space using
the X Color Management spec.

bugs expected ;-)