* [core]: fix DnD
Still after dropping a file name the GL view is not aware of mouse move
events. Only after leaving the window the mouse move events become visible
again. Somehow gets the Fl_Box::handle(event) call in icc_betrachter.cxx in
the way and the dndCommes() workaround is not complete.

tested on oS-12.2 with system FLTK-1.3.0