* [core]: compensate for high DPI displaying
It is tested on a Linux XRandR system. Appears to be completed.

FLTK's Fl::screen_dpi() function looks like reporting always 96 DPI.
Thus a workaround for with using XRandR calls is implemented. A bug
is reported as STR #3273 to http://www.fltk.org .
Only the first screen is choosen for DPI detection and scaling. Ideally
the actual window's screen is used for better scaling.